Narrow Subprocess guard in skit and split subcommands into their own files [skip ci]
Drop the file-level `#if canImport(Subprocess)` around the entire Skit
type and the separate SkitStub `@main`. The single `@main Skit` now
covers all platforms; only the body of `Run.run()` is guarded, so
`skit parse` works without a Subprocess backend. Each subcommand lives
in its own extension file (Skit+Run.swift, Skit+Parse.swift), and Run's
options shed their verbose `discussion:` text in favor of ArgumentParser's
default `--help` rendering.
